+++ public/ace/snippets/qml.js
@@ -1,5 +1,5 @@
-define("ace/snippets/csp",["require","exports","module"],function(e,t,n){"use strict";t.snippetText=undefined,t.scope="csp"});                (function() {
-                    window.require(["ace/snippets/csp"], function(m) {
+define("ace/snippets/qml",["require","exports","module"],function(e,t,n){"use strict";t.snippetText=undefined,t.scope="qml"});                (function() {
+                    window.require(["ace/snippets/qml"], function(m) {
                         if (typeof module == "object" && typeof exports == "object" && module) {
                             module.exports = m;
                         }